+++ imports/api/cdp/cdp.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const CDPCollection = new Mongo.Collection('cdp');