+++ imports/api/hard/hard.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const HARDCollection = new Mongo.Collection('hard');