+++ imports/ui/cdp/List.jsx
@@ -7,6 +7,7 @@ import TimeStamp from '../components/TimeStamp.jsx';
 import Coin from '/both/utils/coins.js'
 import Pagination from "react-js-pagination";
 
+
 const T = i18n.createComponent();
 let minCollateralRatio = 0;
 
@@ -55,27 +56,19 @@ export default class List extends Component {
     }
 
     getCDPList = () => {
-        Meteor.call('cdp.getCDPList', this.props.collateralType, (error, result) => {
-            if(result){
+        Meteor.call('cdp.fetchList', (error, result) => {
+            if (result) {
                 this.setState({
-                    cdpList: result?.length > 1 ? result.map((cdpList, i) => {
+                    cdpList: result?.CDPList[this.props.collateralType]?.length > 1 ? result.CDPList[this.props.collateralType].map((cdpList, i) => {
                         return <CDPRow key={i} index={i} cdpList={cdpList} />
-                    }) : <CDPRow cdpList={result[0]} />,
+                    }) : result?.CDPList[this.props.collateralType]?.length === 0 ? <CDPRow cdpList={result?.CDPList[this.props.collateralType][0]} /> : null,
                     pagesCount: Math.ceil(result.length / this.state.pageSize),
                     currentPage: 1,
                     loading: false,
                     noActiveCDP: false
                 })
             }
-            if(result === null){
-                this.setState({
-                    cdpList: null,
-                    pagesCount: 0,
-                    loading: false,
-                    noActiveCDP: true
-                })
-            }
-            if(error) {
+            if (error) {
                 this.setState({
                     cdpList: undefined,
                     pagesCount: 0,
@@ -87,7 +80,7 @@ export default class List extends Component {
 
 
     getMinCollateralRatio = () => {
-        Meteor.call('cdp.getCDPParams', (error, result) => {
+        Meteor.call('cdp.fetchParameters', (error, result) => {
             if (error) {
                 console.warn(error);
                 this.setState({