+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-# [Unreleased]
+## [UNRELEASED]
 * [#536] Added missing `/` in staking parameters URL 
+* [#535] Stored CDP, HARD, incentive and auctions details in db
 
 # kava-7-v1.1
 * Update npm packages